requirements.txt
mcp
requests
PyYAML
